test(domain): co-locate in_memory_store and card command tests with their submodules (KAN-762)
