+++ crates/kanban-domain/src/commands/card/positioning.rs
@@ -83,3 +83,93 @@ impl CompactColumnPositions {
         Ok(commands)
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::commands::test_helpers::TestContext;
+
+    #[test]
+    fn test_move_card_not_found_returns_error() {
+        let tc = TestContext::new();
+        let column = crate::Column::new(Uuid::new_v4(), "Col", 0);
+        let column_id = column.id;
+        tc.store.upsert_column(column).unwrap();
+        let context = tc.as_command_context();
+        let cmd = MoveCard {
+            card_id: Uuid::new_v4(),
+            new_column_id: column_id,
+            new_position: 0,
+        };
+        let result = cmd.execute(&context);
+        assert!(result.unwrap_err().is_not_found());
+    }
+
+    #[test]
+    fn test_move_card_column_not_found_returns_error() {
+        let tc = TestContext::new();
+        let mut board = crate::Board::new("Test", Some("TST"));
+        let card = crate::Card::new(&mut board, Uuid::new_v4(), "Card", 0);
+        let card_id = card.id;
+        tc.store.upsert_card(card).unwrap();
+        let context = tc.as_command_context();
+        let cmd = MoveCard {
+            card_id,
+            new_column_id: Uuid::new_v4(),
+            new_position: 0,
+        };
+        let result = cmd.execute(&context);
+        assert!(result.unwrap_err().is_not_found());
+    }
+
+    #[test]
+    fn test_move_card_exceeding_wip_limit_returns_error() {
+        let tc = TestContext::new();
+        let mut board = crate::Board::new("Test", Some("TST"));
+        let src_col = crate::Column::new(board.id, "Source", 0);
+        let mut dst_col = crate::Column::new(board.id, "Dest", 1);
+        dst_col.wip_limit = Some(1);
+        let dst_id = dst_col.id;
+        let existing = crate::Card::new(&mut board, dst_id, "Existing", 0);
+        let mover = crate::Card::new(&mut board, src_col.id, "Mover", 0);
+        let mover_id = mover.id;
+        tc.store.upsert_board(board).unwrap();
+        tc.store.upsert_column(src_col).unwrap();
+        tc.store.upsert_column(dst_col).unwrap();
+        tc.store.upsert_card(existing).unwrap();
+        tc.store.upsert_card(mover).unwrap();
+
+        let context = tc.as_command_context();
+        let cmd = MoveCard {
+            card_id: mover_id,
+            new_column_id: dst_id,
+            new_position: 1,
+        };
+        let result = cmd.execute(&context);
+        assert!(result.unwrap_err().is_wip_limit_exceeded());
+    }
+
+    #[test]
+    fn test_compact_column_positions_makes_sequential() {
+        let tc = TestContext::new();
+        let mut board = crate::Board::new("B", Some("TST"));
+        let col = crate::Column::new(board.id, "Col", 0);
+        let column_id = col.id;
+        let mut card1 = crate::Card::new(&mut board, column_id, "C1", 0);
+        card1.position = 0;
+        let mut card2 = crate::Card::new(&mut board, column_id, "C2", 5);
+        card2.position = 5;
+        tc.store.upsert_board(board).unwrap();
+        tc.store.upsert_column(col).unwrap();
+        tc.store.upsert_card(card1).unwrap();
+        tc.store.upsert_card(card2).unwrap();
+
+        let context = tc.as_command_context();
+        let cmd = CompactColumnPositions { column_id };
+        cmd.execute(&context).unwrap();
+
+        let cards = tc.store.list_cards_by_column(column_id).unwrap();
+        assert_eq!(cards[0].position, 0);
+        assert_eq!(cards[1].position, 1);
+    }
+}
