+++ crates/kanban-domain/src/commands/card/tests/mod.rs
@@ -1,0 +1,3 @@
+mod archive_update;
+mod create;
+mod restore;
