+++ crates/kanban-domain/src/in_memory_store/columns.rs
@@ -46,3 +46,57 @@ impl InMemoryStore {
         Ok(())
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::data_store::DataStore;
+    use crate::in_memory_store::test_support::{make_board, make_column};
+
+    #[test]
+    fn test_upsert_and_get_column() {
+        let store = InMemoryStore::new();
+        let board = make_board("B");
+        let col = make_column(board.id, "Col", 0);
+        let col_id = col.id;
+        store.upsert_column(col.clone()).unwrap();
+
+        let fetched = store.get_column(col_id).unwrap().unwrap();
+        assert_eq!(fetched.id, col_id);
+        assert_eq!(fetched.name, "Col");
+    }
+
+    #[test]
+    fn test_list_columns_by_board_filters_correctly() {
+        let store = InMemoryStore::new();
+        let board1 = make_board("B1");
+        let board2 = make_board("B2");
+        let col1 = make_column(board1.id, "C1", 0);
+        let col2 = make_column(board1.id, "C2", 1);
+        let col3 = make_column(board2.id, "C3", 0);
+        store.upsert_column(col1).unwrap();
+        store.upsert_column(col2).unwrap();
+        store.upsert_column(col3).unwrap();
+
+        let cols = store.list_columns_by_board(board1.id).unwrap();
+        assert_eq!(cols.len(), 2);
+        assert!(cols.iter().all(|c| c.board_id == board1.id));
+    }
+
+    #[test]
+    fn test_delete_columns_by_board() {
+        let store = InMemoryStore::new();
+        let board1 = make_board("B1");
+        let board2 = make_board("B2");
+        let col1 = make_column(board1.id, "C1", 0);
+        let col2 = make_column(board2.id, "C2", 0);
+        let col2_id = col2.id;
+        store.upsert_column(col1).unwrap();
+        store.upsert_column(col2).unwrap();
+
+        store.delete_columns_by_board(board1.id).unwrap();
+
+        assert!(store.list_columns_by_board(board1.id).unwrap().is_empty());
+        assert!(store.get_column(col2_id).unwrap().is_some());
+    }
+}
